requirements.txt
fastapi
uvicorn
python-dotenv
requests
amadeus
crewai
crewai-tools
pyyaml
